feat: Form component with built-in functionality
This is a prototype. It's not finished yet. I want to build a Form
component with built-in functionalities common to many forms. Such as
closing the form when the users press the Esc key.Right now, I'm not
clear on what features I want to add. I don't know if the forms for
login/register and for add/modify tasks info are so different that I'll
have to make the component polymorphic to support both.